Added boss and baby elf under enemy class
