+++ goblin.py
@@ -1,27 +1,10 @@
 import random
+from enemy import Enemy
 
-class Goblin:
-    """
-    This is our goblin blueprint 
+class Goblin(Enemy):
     
-    Attributes:
-        name: Awe, it has a name? How cute!
-        health: The current health value 
-        attack_power: How much health will be drained from opponent if hit
-    """
-    def __init__(self, name):
-        self.name = name
-        self.health = 100
-        self.attack_power = random.randint(5, 15)
-
-    def attack(self):
-        return random.randint(1, self.attack_power)
-
-    def take_damage(self, damage):
-        self.health -= damage
-        if self.health < 0:
-            self.health = 0
-        print(f"{self.name} takes {damage} damage. Health is now {self.health}.")
-
-    def is_alive(self):
-        return self.health > 0
+    def __init__(self, name, color):
+        super().__init__(name)
+        self.color = color
+        self.health = 50
+
